# openapi.yaml
openapi: 3.0.3
info:
  title: Robot Scheduler API
  description: |
    机器人调度系统后端服务接口文档。
    包含任务管理、机器人管理、调度触发、日志查询、LLM 交互、SLAM 地图管理等模块。
  version: 1.0.0
  contact:
    name: Robot Scheduler Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
  /api/v1/tasks:
    post:
      summary: 创建任务
      description: 创建新任务并自动触发调度
      tags:
        - 任务管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultTaskCreateResponse'
    get:
      summary: 获取任务列表
      description: 支持按状态和机器人ID筛选
      tags:
        - 任务管理
      parameters:
        - name: status
          in: query
          schema:
            type: string
            example: RUNNING
          description: 任务状态筛选
        - name: robotId
          in: query
          schema:
            type: string
          description: 机器人ID筛选
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultTaskListResponse'
  /api/v1/tasks/{taskId}:
    get:
      summary: 获取任务详情
      tags:
        - 任务管理
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultTaskDetailResponse'
  /api/v1/tasks/{taskId}/status:
    patch:
      summary: 更新任务状态
      description: |
        更新任务状态。状态值使用英文：QUEUED / RUNNING / SUCCESS / FAILED。
        更新为 SUCCESS 或 FAILED 时自动记录日志。
      tags:
        - 任务管理
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultTaskStatusResponse'
...
components:
  schemas:
    Result:
      type: object
      properties:
        code:
          type: integer
          description: 状态码
          example: 200
        message:
          type: string
          description: 响应信息
          example: "操作成功"
        data:
          description: 业务数据
...
    TaskCreateRequest:
      type: object
      properties:
        robotId:
          type: string
          description: 指定机器人ID（可选）
        commandType:
          type: string
          description: 命令类型
          example: NAVIGATE
        priority:
          type: integer
          description: 优先级（默认3）
          example: 3
        params:
          type: object
          description: 任务参数
          additionalProperties: true
          example:
            target: "A1"
            speed: 0.5
    TaskCreateResponse:
      type: object
      properties:
        taskId:
          type: string
        robotId:
          type: string
        robotCode:
          type: string
        commandType:
          type: string
        priority:
          type: integer
        status:
          type: string
        createdAt:
          type: integer
          format: int64
          description: 时间戳（毫秒）
        params:
          type: object
    ResultTaskCreateResponse:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TaskCreateResponse'
    TaskListItem:
      type: object
      properties:
        taskId:
          type: string
        robotId:
          type: string
        robotCode:
          type: string
        commandType:
          type: string
        priority:
          type: integer
        status:
          type: string
        createdAt:
          type: integer
          format: int64
        params:
          type: string
          description: JSON字符串
    ResultTaskListResponse:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/TaskListItem'
    TaskDetailResponse:
      type: object
      properties:
        taskId:
          type: string
        robotId:
          type: string
        robotCode:
          type: string
        commandType:
          type: string
        priority:
          type: integer
        status:
          type: string
        createdAt:
          type: integer
          format: int64
        params:
          type: string
    ResultTaskDetailResponse:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TaskDetailResponse'
    TaskStatusUpdateRequest:
      type: object
      properties:
        status:
          type: string
          description: 新状态（RUNNING / SUCCESS / FAILED）
          example: SUCCESS
        reason:
          type: string
          description: 状态变更原因（失败时必填）
    TaskStatusResponse:
      type: object
      properties:
        taskId:
          type: string
        status:
          type: string
    ResultTaskStatusResponse:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TaskStatusResponse'
...